Create Todo API client to wrap the interaction with the backend API
