--- src/api/healthcheck.ts
@@ -1,11 +1,0 @@
-import axios from 'axios'
-
-export const isApiHealthy = async (apiBaseUrl: string): Promise<boolean> => {
-  const instance = axios.create({
-    baseURL: apiBaseUrl
-  })
-  return instance
-    .get('/health')
-    .then(response => response.data.status)
-    .then(text => text === 'ok')
-}
