+++ src/api/todos.pact.ts
@@ -1,9 +1,12 @@
-import { eachLike, somethingLike } from '@pact-foundation/pact/dsl/matchers'
+import {
+  term,
+  eachLike,
+  somethingLike
+} from '@pact-foundation/pact/dsl/matchers'
 import { setupTodoApiIntegration } from '@todo/support/pact'
-import { getTodos } from './todos'
 
 describe('Todos API pact test', () => {
-  const { provider } = setupTodoApiIntegration()
+  const { provider, createClient } = setupTodoApiIntegration()
 
   it('fetches list of todos', async () => {
     await provider.addInteraction({
@@ -24,7 +27,33 @@ describe('Todos API pact test', () => {
       }
     })
 
-    const todos = await getTodos(provider.mockService.baseUrl)
+    const todos = await createClient().getTodos()
     expect(todos).toBeTruthy()
+  })
+
+  it('checks API status', async () => {
+    await provider.addInteraction({
+      state: 'application is healthy',
+      uponReceiving: 'healthcheck',
+      withRequest: {
+        method: 'GET',
+        path: '/health'
+      },
+      willRespondWith: {
+        status: 200,
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body: {
+          status: term({
+            matcher: 'ok|fail',
+            generate: 'ok'
+          })
+        }
+      }
+    })
+
+    const healthResponse = await createClient().isApiHealthy()
+    expect(healthResponse).toBeTruthy()
   })
 })
