--- src/api/todos.ts
@@ -1,12 +1,0 @@
-import axios from 'axios'
-
-interface Todo {
-  todo: string
-}
-
-export const getTodos = async (apiBaseUrl: string): Promise<Todo[]> => {
-  const instance = axios.create({
-    baseURL: apiBaseUrl
-  })
-  return instance.get('/todos').then(response => response.data)
-}
